Refactor Map to a component
